Rename Index.vue to FrontPage.vue
