+++ client/src/views/FrontPage.vue
